pyproject.toml
[project]
name = "harbor"
version = "0.1.0"
description = "Agent Platform Management — Registry, Discovery, and Governance for AI Agents on AWS Bedrock AgentCore"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "PyJWT[crypto]>=2.9.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
cli = [
    "click>=8.0",
    "httpx>=0.28.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
    "mypy>=1.13",
    "moto[dynamodb]>=5.0",
    "httpx>=0.28.0",
    "click>=8.0",
]

[project.scripts]
harbor = "harbor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
